refactor(gui): extract voice/assets tests to sibling file
Move the single trailing `#[cfg(test)] mod tests` block out of
`primer-gui/src/voice/assets.rs` (714 lines) into a sibling
`src/voice/assets/tests.rs`. Pure verbatim relocation — no behaviour
change, no incidental production-code fixes.

`assets` is gated behind `#[cfg(feature = "speech")]` in
`voice/mod.rs`, and `assets.rs` is a non-root module, so its submodule
resolves to `src/voice/assets/tests.rs` (new `assets/` dir). The tests
are a descendant module, so `use super::*` still reaches every private
parent item — zero visibility churn.

assets.rs drops 714 -> 274 lines (-440, now under the ~500 line
guideline). The only parent-diff addition is `mod tests;`.

Verification:
- cargo test -p primer-gui --features speech --lib: 253 passed
  (identical to baseline)
- cargo test -p primer-gui --lib (default): 205 passed
- cargo clippy -p primer-gui {--features speech,} --all-targets
  -D warnings: clean
- cargo fmt --all --check: clean
- cargo test --workspace: 51 ok blocks, 0 failures
